feat: 파트별 커리큘럼 섹션 구현
